+++ src/admin-user-list/user-list.controller.js
@@ -103,9 +103,9 @@
          * Holds options for sorting user list.
          */
         vm.options = {
-            firstName: 'adminUserList.firstName',
-            lastName: 'adminUserList.lastName',
-            username: 'adminUserList.username'
+            'adminUserList.firstName': ['firstName'],
+            'adminUserList.lastName': ['lastName'],
+            'adminUserList.username': ['username']
         };
 
         /**