+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Improvements:
 * [OLMIS-5235](https://openlmis.atlassian.net/browse/OLMIS-5235): Improved performance of login:
 ** Made the requests when login concurrent (where possible).
 ** Minimal Facilities are now cached in the LocalDatabase instead of the LocalStorage.
+* [OLMIS-5409](https://openlmis.atlassian.net/browse/OLMIS-5409): Updated ui-components to version 7.0.0.
 
 Bug fixes:
 * [OLMIS-5235](https://openlmis.atlassian.net/browse/OLMIS-5235): FacilityProgramCacheService will now pass user ID when fetching user programs.